rename section labeling module
