+++ frontend/javascripts/viewer/model/volumetracing/section_labeling.ts
